refactor : project refactoring

+++ src/main/java/com/eunbinlib/api/domain/request/PostSearch.java
@@ -28,6 +28,7 @@ public class PostSearch {
     }
 
     public long getLimit() {
-        return (long) min(size, MAX_SIZE);
+        return min(size, MAX_SIZE);
     }
+
 }
